+++ core/class/AbeilleCmdPrepare.class.php
@@ -473,24 +473,24 @@
             //         "action"                => $actionId,
             //     );
             //     break;
-            case "OnOffGroupBroadcast":
-                if ($msg == "On") {
-                    $actionId = "01";
-                } elseif ($msg == "Off") {
-                    $actionId = "00";
-                } elseif ($msg == "Toggle") {
-                    $actionId = "02";
-                }
-                $Command = array(
-                    "onoff"                 => "1",
-                    "addressMode"           => "04",
-                    "priority"              => $priority,
-                    "dest"                  => $dest,
-                    "address"               => $address,
-                    "destinationEndpoint"   => "01", // Set but not send on radio
-                    "action"                => $actionId,
-                );
-                break;
+            // case "OnOffGroupBroadcast": // Obsolete. Replaced by 'cmd-0006'
+            //     if ($msg == "On") {
+            //         $actionId = "01";
+            //     } elseif ($msg == "Off") {
+            //         $actionId = "00";
+            //     } elseif ($msg == "Toggle") {
+            //         $actionId = "02";
+            //     }
+            //     $Command = array(
+            //         "onoff"                 => "1",
+            //         "addressMode"           => "04",
+            //         "priority"              => $priority,
+            //         "dest"                  => $dest,
+            //         "address"               => $address,
+            //         "destinationEndpoint"   => "01", // Set but not send on radio
+            //         "action"                => $actionId,
+            //     );
+            //     break;
             case "OnOffGroupTimed":
                 $fields = preg_split("/[=&]+/", $msg);
                 if (count($fields) > 1) {